samza-test/src/main/python/requirements.txt
zopkio
requests
kafka-python
